copy over colin's S3 fix; pull Ec2port out to its own file
