+++ pulumi/infra/dgraph_cluster.py
@@ -4,7 +4,8 @@ from typing import Any, Optional
 import pulumi_aws as aws
 from infra.bucket import Bucket
 from infra.config import DEPLOYMENT_NAME
-from infra.swarm import Ec2Port, Swarm
+from infra.swarm import Swarm
+from infra.ec2 import Ec2Port
 
 import pulumi
 from pulumi.output import Output
